requirements.txt
Pillow~=10.4.0
Telethon~=1.36.0
requests~=2.32.3